+++ src/models/productDao.js
@@ -1,31 +1,9 @@
-const {DataSource} = require("typeorm");
-const path = require("path");
-const envPath = path.join(__dirname, "../utils","env")
-const dotenv = require("dotenv");
-const { json } = require("body-parser");
-dotenv.config({path: envPath});
-
-const appDataSoure = new DataSource({
-    type: process.env.TYPEORM_CONNECTION,
-    host: process.env.TYPEORM_HOST,
-    port: process.env.TYPEORM_PORT,
-    username: process.env.TYPEORM_USERNAME,
-    password: process.env.TYPEORM_PASSWORD,
-    database: process.env.TYPEORM_DATABASE,
-})
-
-appDataSoure.initialize()
-.then(() => {
-    console.log("Data Source has been initialize");
-}).catch((err) => {
-    console.err("Error occurred during Data Source initialization", err)
-})
-
+const database = require("../utils/database")
 
 // 유저 정보 조회
 const selectUserInfo = async(userId, email) => {
     try{
-        const result = await appDataSoure.query(
+        const result = await database.appDataSoure.query(
             `
             select * from users
             where id =? and email =?
@@ -43,7 +21,7 @@ const selectUserInfo = async(userId, email) => {
 
 // 장바구니에 담긴 데이터 확인
 const selectCart = async(cartId) => {
-    const result = await appDataSoure.query(
+    const result = await database.appDataSoure.query(
         `
         select * from ShoppingItems
         where = ?
@@ -56,7 +34,7 @@ const selectCart = async(cartId) => {
 // 유저의 point 변경
 const cost = async(userPoint, userId) => {
     try{
-        const result = await appDataSoure.query(
+        const result = await database.appDataSoure.query(
             `
             update users set point = ?
             where id = ?;
@@ -76,7 +54,7 @@ const cost = async(userPoint, userId) => {
 // 결제 정보를 담기 위한 cart 정보 불러오기
 const cartList = async(userId) => {
     try{
-        const result = appDataSoure.query(
+        const result = database.appDataSoure.query(
             `
             select * from ShoppingItems
             where userId = ?
@@ -98,7 +76,7 @@ const cartList = async(userId) => {
 // 결제 정보 담기
 const payment = async(cartId, totalPrice) => {
     try{
-        const result = await appDataSoure.query(
+        const result = await database.appDataSoure.query(
             `
             insert into Payments (shoppingItemsId, totalPrice)
             values (?, ?)
@@ -119,7 +97,7 @@ const payment = async(cartId, totalPrice) => {
 // 결제 상태 업데이트
 const updateStatus = async(payStatus, userId) => {
     try{
-        const result = await appDataSoure.query(
+        const result = await database.appDataSoure.query(
             `
             update ShoppingItems
             set status = ?
